utilisation generalisee de autoriser implementation de autoriser_form autoriser_compat pour permettre un support de la 1.9.1
git-svn-id: svn://zone.spip.org/spip-zone/_plugins_@7262 ac52e18a-acf5-0310-9fe8-c4428f23b10a

+++ _stable_/forms/forms2/action/forms_supprime.php
@@ -19,14 +19,16 @@ function action_forms_supprime(){
 	$id_auteur = $auteur_session['id_auteur'];
 	$redirect = _request('redirect');
 	if ($redirect==NULL) $redirect="";
-	include_spip("inc/securiser_action");
+	if (!include_spip("inc/securiser_action"))
+		include_spip("inc/actions");
 	if (verifier_action_auteur("forms_supprime-$id_form",$hash,$id_auteur)==TRUE) {
-		//echo('ok');
-		//if (Forms_form_administrable($id_form)) {
+		if (!include_spip('inc/autoriser'))
+			include_spip('inc/autoriser_compat');
+		if (autoriser('supprimer','form',$id_form)){
 			$result = spip_query("DELETE FROM spip_forms WHERE id_form="._q($id_form));
 			$result = spip_query("DELETE FROM spip_forms_champs WHERE id_form="._q($id_form));
 			$result = spip_query("DELETE FROM spip_forms_champs_choix WHERE id_form="._q($id_form));
-		//}
+		}
 	}
 	redirige_par_entete(str_replace("&amp;","&",urldecode($redirect)));
 }
